Refactor into worker.ui
